Move XmlMiddlewareBuilder to new middleware-config module
The XmlMiddlewareBuilder features are useful for many modules. Moving
the implementation to the new module simplifies its reuse.
