+++ application-common/middleware-config/src/main/java/it/bz/opendatahub/alpinebitsserver/application/common/middleware/config/XmlMiddlewareBuilder.java
@@ -4,7 +4,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package it.bz.opendatahub.alpinebitsserver.odh.inventory.middleware;
+package it.bz.opendatahub.alpinebitsserver.application.common.middleware.config;
 
 import it.bz.opendatahub.alpinebits.middleware.Key;
 import it.bz.opendatahub.alpinebits.middleware.Middleware;
