+++ odh-inventory/impl/src/main/java/it/bz/opendatahub/alpinebitsserver/odh/inventory/middleware/InventoryPullMiddlewareBuilder.java
@@ -16,6 +16,7 @@ import it.bz.opendatahub.alpinebits.middleware.Key;
 import it.bz.opendatahub.alpinebits.middleware.Middleware;
 import it.bz.opendatahub.alpinebits.xml.schema.v_2017_10.OTAHotelDescriptiveInfoRQ;
 import it.bz.opendatahub.alpinebits.xml.schema.v_2017_10.OTAHotelDescriptiveInfoRS;
+import it.bz.opendatahub.alpinebitsserver.application.common.middleware.config.XmlMiddlewareBuilder;
 
 import javax.xml.bind.JAXBException;
 import java.util.Arrays;
